更精确的fps

+++ android-godeye/src/main/java/cn/hikyson/godeye/core/internal/modules/fps/FpsInfo.java
@@ -4,10 +4,10 @@ package cn.hikyson.godeye.core.internal.modules.fps;
  * Created by kysonchao on 2017/11/23.
  */
 public class FpsInfo {
-    public double currentFps;
-    public double systemFps;
+    public int currentFps;
+    public int systemFps;
 
-    public FpsInfo(double currentFps, double systemFps) {
+    public FpsInfo(int currentFps, int systemFps) {
         this.currentFps = currentFps;
         this.systemFps = systemFps;
     }
